+++ src/Models/CustomAttributes/NumberCustomAttribute.php
@@ -71,9 +71,4 @@ class NumberCustomAttribute extends CustomAttribute
         return $this->configureField($field, $owner);
     }
 
-    #[Override]
-    protected function getFingerprintData(): array
-    {
-        return [$this->integerOnly, $this->min, $this->max];
-    }
 }